Tweak perf tradeoff for `getMax`
Prefer repeatedly comparing integers to allocating an extra array: this
code will tend to be in a very, *very* hot path. Numeric comparisons
are about the cheapest possible operation, while allocating arrays is
very much *not* cheap. The previous implementation here allocated an
extra array for the `map` call and then iterated it into an `arguments`
array-like for the `Math.max` call; the new implementation simply does
a single iteration of the `tags` array and never allocates. We could
also use `Array.prototype.reduce` here, but I expect the `for...of` is
actually clearer and more obvious about its intent.