Make `Derived` a standalone class, not a `Signal` subclass
At this point, `Derived` is not really making use of the fact that it
is a subclass, and avoiding polymorphism tends to help performance in
JS VMs. As part of this:

- Introduce a `#tag` property and use `createTag` directly.

- For now at least, explicitly use `===` equality for comparison of the
  previous and current values. (This was happening implicitly already
  since the `super` call in the `constructor` was not passing along a
  custom equality check!). If/when we introduce the ability for users
  to receive the previous value in a `Derive` callback, they will be
  able to do their own equality check *there*.

This allows for a few similar changes to `Signal`, which no longer
needs to expose any of its internal properties as `protected` for the
sake of subclassing.

Additionally, since the `constructor`s for `Signal` and `Derived` will
always constrain `T`, we can rely on inference to drive the type of
actual concrete items, so we do not *need* the `T = unknown`, and
reliance on default type parameters tends to lead end users into bad
places where they think it is causing constraints that it is not, and
simultaneously *avoiding* inference where it would be helpful. The cost
is small: being explicit about not caring about the wrapped type where
one doesn't care.