Inline the `Tag` type onto `Signal` and `Derived`
This keeps the same basic contract as we had previously, but it does it
in a way which avoids allocating an extra `Tag` object for every single
`Signal` or `Derived` instance. The idea is: introduce the revision
count onto the items themselves, and define an interface which makes
that contract explicit. Additionally, so that we don't accidentally
compare *other* `number` values to tags, provide an opaque "new type"
definition for `Tag`, which is only ever cast in the `tag` and `state`
modules, so that consumers can treat it as a discrete type, but things
like `Math.max` still work.

Net, this means we may keep around `Signal` and `Derived` one extra
cycle in cases where they are dependencies of other `Derived` or
`Effect` instances, but in exchange we get to avoid allocating fully
half the objects we were allocating otherwise: even if small, that can
add up when there are many of these.