Fix type error caused by variance on `Signal.isEqual`
Define the public interface of `Signal` to exclude its `isEqual` field.
As a private field, consumers cannot use it *anyway*, and `Signal` is
not allowed to be subclassed. Providing the public type for it in terms
of an `interface` which `Omit`s the `isEqual` field simultaneously
prevents subclassing *and* dodges that variance issue.