Eliminate the need for casts in `createSignal()`
By explicitly distinguishing between non-nullable and nullable creation
paths in the overloads, and then explicitly checking `== null` in the
constructor call, we give TS enough info to eliminate the need for the
casts. This also requires propagating that back to `useSignal()`, which
is a mild annoyance, but provides *much* better type safe behavior.